getplans, searchplans, template
