+++ client/src/api/axios.ts
@@ -25,16 +25,8 @@ async function GetPlans(projectId:any){
   })
 }
 
-async function CreateNewTestSuite(testSuiteDetails: any, project_id:any){
-  await AuthClient.post(`/test_suites/${project_id}/`)
-  .then(response=>{
-  })
-  .catch(error=>{
-    return error
-  })
-}
-async function GetTestSuites(project_id:any){
-  await AuthClient.get(`/test_suites/${project_id}/`)
+async function SearchPlans(project_id:any,key_word:any){
+  await AuthClient.post(`/test_plan/${project_id}/search/${key_word}/`)
   .then(response=>{
     return response.data
   })
@@ -43,14 +35,4 @@ async function GetTestSuites(project_id:any){
   })
 }
 
-async function SearchSuite(project_id:any,key_word:any){
-  await AuthClient.get(`/test_suites/${project_id}/search/${key_word}/`)
-  .then(response=>{
-    return response.data
-  })
-  .catch(error=>{
-    return error
-  })
-}
-
-export default{ AuthClient, BaseClient,GetPlans, CreateNewTestSuite, GetTestSuites,SearchSuite };
+export default{ AuthClient, BaseClient,GetPlans,SearchPlans };
