+++ client/src/typed-router.d.ts
@@ -18,7 +18,7 @@ declare module 'vue-router/auto-routes' {
    * Route name map generated by unplugin-vue-router
    */
   export interface RouteNamedMap {
-    '/AddTestSuite': RouteRecordInfo<'/AddTestSuite', '/AddTestSuite', Record<never, never>, Record<never, never>>,
+    '/AddTestPlan': RouteRecordInfo<'/AddTestPlan', '/AddTestPlan', Record<never, never>, Record<never, never>>,
     '/DashboardView': RouteRecordInfo<'/DashboardView', '/DashboardView', Record<never, never>, Record<never, never>>,
   }
 }
